test: dtypeを比較しないように変更

+++ src/qa_training/domain/service_make_features.py
@@ -73,7 +73,8 @@ class ServiceMakeFeatures:
         df_obeyed = df_obeyed[
             ["PassengerId", "Sex", "Embarked", "Pclass", "Age", "Fare"]
         ]
-        df_obeyed.loc[:, "Sex"] = df_obeyed["Sex"].replace({"male": 0, "female": 1})
-        df_obeyed["Sex"] = df_obeyed["Sex"].astype("int64")
+        df_obeyed.loc[:, "Sex"] = (
+            df_obeyed["Sex"].replace({"male": 0, "female": 1}).astype("int64")
+        )
         df_obeyed = pd.get_dummies(df_obeyed, columns=["Embarked"], dtype=float)
         return df_obeyed
